Small refactoring. Reformulate GrainFactory and ServiceFactopry. Start implementing activatiion of the grain within searching for the grain.
