--- src/main/scala/org/orleans/common/AccountGrain.scala
@@ -1,8 +1,0 @@
-package org.orleans.common
-
-
-class AccountGrain extends Grain{
-    def store(): Unit = {
-      println("Executing store method for account grain")
-    }
-}
