--- src/main/scala/org/orleans/common/Grain.scala
@@ -1,5 +1,0 @@
-package org.orleans.common
-
-trait Grain extends Serializable {
-  def store()
-}
