+++ src/main/scala/org/orleans/silo/Services/Grain/Grain.scala
@@ -1,0 +1,5 @@
+package org.orleans.silo.Services.Grain
+
+trait Grain extends Serializable {
+  def store()
+}
